+++ vkconfig_core/setting_type.h
@@ -42,7 +42,7 @@ enum SettingType {  // Enum value can't be changed
 enum { SETTING_COUNT = SETTING_LAST - SETTING_FIRST + 1 };
 
 SettingType GetSettingType(const char* token);
-const char* GetSettingToken(SettingType type);
+const char* GetSettingTypeToken(SettingType type);
 
 inline bool IsEnum(SettingType type) {
     assert(type >= SETTING_FIRST && type <= SETTING_LAST);
@@ -55,3 +55,17 @@ inline bool IsPath(SettingType type) {
 
     return type == SETTING_SAVE_FILE || type == SETTING_LOAD_FILE || type == SETTING_SAVE_FOLDER;
 }
+
+enum SettingView {
+    SETTING_VIEW_STANDARD = 0,
+    SETTING_VIEW_ADVANCED,
+    SETTING_VIEW_HIDDEN,
+
+    SETTING_VIEW_FIRST = SETTING_VIEW_STANDARD,
+    SETTING_VIEW_LAST = SETTING_VIEW_HIDDEN
+};
+
+enum { SETTING_VIEW_COUNT = SETTING_VIEW_LAST - SETTING_VIEW_FIRST + 1 };
+
+SettingView GetSettingView(const char* token);
+const char* GetSettingViewToken(SettingView state);