+++ vkconfig/CHANGELOG.md
@@ -28,6 +28,9 @@
 ### Improvements:
 - Add display of layer execution order in the setting tree #1390
 
+### Fixes:
+- Hide 'Callback' from validation layer 'Debug Action' debug action that is NOOP
+
 ## [Vulkan Configurator 2.2.0 for Vulkan SDK 1.2.170.0](https://github.com/LunarG/VulkanTools/tree/sdk-1.2.170.0) - February 2021
 
 ### Features: